requirements.txt
python-dotenv
zhipuai
rich
pyyaml